rename loopcomplement to loopanti
