+++ pkg/sql/colexec/loopanti/loopanti_test.go
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package loopcomplement
+package loopanti
 
 import (
 	"bytes"
