+++ pkg/sql/colexec/loopanti/types.go
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package loopcomplement
+package loopanti
 
 import (
 	"github.com/matrixorigin/matrixone/pkg/container/batch"
